refactor: Remove redundant mock implementations from tests
- Delete the mockSearchManager implementation from the httpd module test file to streamline test code.
- Consolidate mock configurations in the api module test file for better organization and clarity.
- Ensure that all necessary mock expectations are maintained while improving overall test readability.
- Enhance consistency in mock usage across test files, aligning with recent refactoring efforts.
